+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/Entities.hs
@@ -20,20 +20,35 @@ module Cardano.Ledger.Dijkstra.Rules.Entities (
   EntitiesEvent (..),
   validateWrongNetworkInDirectDeposit,
   validateMissingAccountsInDirectDeposits,
+  checkAccountBalanceIntervals,
 ) where
 
-import Cardano.Ledger.Address (DirectDeposits (..))
+import Cardano.Ledger.Address (DirectDeposits (..), accountAddressCredentialL)
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Coin (Coin)
+import Cardano.Ledger.Compactible (fromCompact)
 import Cardano.Ledger.Conway.Core
+import Cardano.Ledger.Conway.Governance (
+  Committee,
+  GovActionPurpose (..),
+  GovActionState,
+  GovPurposeId,
+ )
 import qualified Cardano.Ledger.Conway.Rules as Conway
 import Cardano.Ledger.Conway.State
 import Cardano.Ledger.Dijkstra.Era (DijkstraEra, ENTITIES)
 import Cardano.Ledger.Dijkstra.Rules.Certs ()
 import Cardano.Ledger.Dijkstra.Rules.GovCert (DijkstraGovCertPredFailure)
-import Cardano.Ledger.Dijkstra.TxBody (DijkstraEraTxBody, directDepositsTxBodyL)
+import Cardano.Ledger.Dijkstra.Scripts (AccountBalanceInterval (..), AccountBalanceIntervals (..))
+import Cardano.Ledger.Dijkstra.TxBody (
+  DijkstraEraTxBody,
+  accountBalanceIntervalsTxBodyL,
+  directDepositsTxBodyL,
+  startingAccountBalanceIntervalsTxBodyL,
+ )
+import Cardano.Ledger.Dijkstra.UTxO (DijkstraEraUTxO (..))
 import Cardano.Ledger.Rules.ValidationMode (Test, runTest)
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
 import Control.DeepSeq (NFData)
@@ -43,29 +58,50 @@ import Data.Map.NonEmpty (NonEmptyMap)
 import qualified Data.Map.NonEmpty as NEM
 import qualified Data.Map.Strict as Map
 import Data.Sequence (Seq)
+import qualified Data.Sequence.Strict as StrictSeq
+import Data.Set (Set)
 import Data.Set.NonEmpty (NonEmptySet)
 import GHC.Generics (Generic)
 import Lens.Micro
 
 data EntitiesEnv era = EntitiesEnv
-  { eePlutusLegacyMode :: Bool
-  , eeCertsEnv :: Conway.CertsEnv era
+  { eeCurrentEpoch :: EpochNo
+  , eePParams :: PParams era
+  , eeCurrentCommittee :: StrictMaybe (Committee era)
+  , eeCommitteeProposals :: Map.Map (GovPurposeId 'CommitteePurpose) (GovActionState era)
+  , eeOriginalAccounts :: Accounts era
   }
   deriving (Generic)
 
-deriving instance (EraPParams era, Eq (Tx TopTx era)) => Eq (EntitiesEnv era)
+deriving instance
+  (EraPParams era, Eq (Committee era), Eq (GovActionState era), Eq (Accounts era)) =>
+  Eq (EntitiesEnv era)
 
-deriving instance (EraPParams era, Show (Tx TopTx era)) => Show (EntitiesEnv era)
+deriving instance
+  (EraPParams era, Show (Committee era), Show (GovActionState era), Show (Accounts era)) =>
+  Show (EntitiesEnv era)
 
-instance (EraPParams era, NFData (Tx TopTx era)) => NFData (EntitiesEnv era)
+instance
+  (EraPParams era, NFData (Committee era), NFData (GovActionState era), NFData (Accounts era)) =>
+  NFData (EntitiesEnv era)
 
-instance EraTx era => EncCBOR (EntitiesEnv era) where
-  encCBOR x@(EntitiesEnv _ _) =
+instance
+  ( EraPParams era
+  , EncCBOR (Committee era)
+  , EncCBOR (GovActionState era)
+  , EncCBOR (Accounts era)
+  ) =>
+  EncCBOR (EntitiesEnv era)
+  where
+  encCBOR x@(EntitiesEnv _ _ _ _ _) =
     let EntitiesEnv {..} = x
      in encode $
           Rec EntitiesEnv
-            !> To eePlutusLegacyMode
-            !> To eeCertsEnv
+            !> To eeCurrentEpoch
+            !> To eePParams
+            !> To eeCurrentCommittee
+            !> To eeCommitteeProposals
+            !> To eeOriginalAccounts
 
 data EntitiesPredFailure era
   = CertsFailure (PredicateFailure (EraRule "CERTS" era))
@@ -83,6 +119,13 @@ data EntitiesPredFailure era
       Network
       -- | Direct-deposit accounts with wrong network id
       (NonEmptySet AccountAddress)
+  | AccountBalanceIntervalsWrongNetwork Network (NonEmptySet AccountAddress)
+  | AccountBalanceIntervalsUnknownAccounts (NonEmptyMap AccountAddress (AccountBalanceInterval era))
+  | AccountBalanceIntervalsMismatch (NonEmptyMap AccountAddress (AccountBalanceInterval era))
+  | StartingAccountBalanceIntervalsWrongNetwork Network (NonEmptySet AccountAddress)
+  | StartingAccountBalanceIntervalsUnknownAccounts
+      (NonEmptyMap AccountAddress (AccountBalanceInterval era))
+  | StartingAccountBalanceIntervalsMismatch (NonEmptyMap AccountAddress (AccountBalanceInterval era))
   deriving (Generic)
 
 deriving stock instance
@@ -113,6 +156,12 @@ instance
       MissingAccountsInDirectDeposits x -> Sum (MissingAccountsInDirectDeposits @era) 4 !> To x
       WrongNetworkInWithdrawals x y -> Sum (WrongNetworkInWithdrawals @era) 5 !> To x !> To y
       WrongNetworkInDirectDeposits x y -> Sum (WrongNetworkInDirectDeposits @era) 6 !> To x !> To y
+      AccountBalanceIntervalsWrongNetwork x y -> Sum (AccountBalanceIntervalsWrongNetwork @era) 7 !> To x !> To y
+      AccountBalanceIntervalsUnknownAccounts x -> Sum (AccountBalanceIntervalsUnknownAccounts @era) 8 !> To x
+      AccountBalanceIntervalsMismatch x -> Sum (AccountBalanceIntervalsMismatch @era) 9 !> To x
+      StartingAccountBalanceIntervalsWrongNetwork x y -> Sum (StartingAccountBalanceIntervalsWrongNetwork @era) 10 !> To x !> To y
+      StartingAccountBalanceIntervalsUnknownAccounts x -> Sum (StartingAccountBalanceIntervalsUnknownAccounts @era) 11 !> To x
+      StartingAccountBalanceIntervalsMismatch x -> Sum (StartingAccountBalanceIntervalsMismatch @era) 12 !> To x
 
 instance
   ( Era era
@@ -128,6 +177,12 @@ instance
     4 -> SumD MissingAccountsInDirectDeposits <! From
     5 -> SumD WrongNetworkInWithdrawals <! From <! From
     6 -> SumD WrongNetworkInDirectDeposits <! From <! From
+    7 -> SumD AccountBalanceIntervalsWrongNetwork <! From <! From
+    8 -> SumD AccountBalanceIntervalsUnknownAccounts <! From
+    9 -> SumD AccountBalanceIntervalsMismatch <! From
+    10 -> SumD StartingAccountBalanceIntervalsWrongNetwork <! From <! From
+    11 -> SumD StartingAccountBalanceIntervalsUnknownAccounts <! From
+    12 -> SumD StartingAccountBalanceIntervalsMismatch <! From
     n -> Invalid n
 
 newtype EntitiesEvent era = CertsEvent (Event (EraRule "CERTS" era))
@@ -170,6 +225,7 @@ instance InjectRuleFailure "ENTITIES" Shelley.ShelleyUtxoPredFailure DijkstraEra
 instance
   ( EraTx era
   , DijkstraEraTxBody era
+  , DijkstraEraUTxO era
   , ConwayEraPParams era
   , ConwayEraCertState era
   , Embed (EraRule "CERTS" era) (ENTITIES era)
@@ -184,7 +240,7 @@ instance
   STS (ENTITIES era)
   where
   type State (ENTITIES era) = CertState era
-  type Signal (ENTITIES era) = Seq (TxCert era)
+  type Signal (ENTITIES era) = StAnnTx TopTx era
   type Environment (ENTITIES era) = EntitiesEnv era
   type BaseM (ENTITIES era) = ShelleyBase
   type PredicateFailure (ENTITIES era) = EntitiesPredFailure era
@@ -195,8 +251,8 @@ instance
 
 dijkstraEntitiesTransition ::
   forall era.
-  ( EraTx era
-  , DijkstraEraTxBody era
+  ( DijkstraEraTxBody era
+  , DijkstraEraUTxO era
   , ConwayEraCertState era
   , Embed (EraRule "CERTS" era) (ENTITIES era)
   , State (EraRule "CERTS" era) ~ CertState era
@@ -209,16 +265,33 @@ dijkstraEntitiesTransition ::
   ) =>
   TransitionRule (ENTITIES era)
 dijkstraEntitiesTransition = do
-  TRC (EntitiesEnv legacyMode certsEnv, certState, certificates) <- judgmentContext
-  let Conway.CertsEnv tx pp curEpoch _committee _committeeProposals = certsEnv
+  TRC (EntitiesEnv curEpoch pp committee committeeProposals originalAccounts, certState, stAnnTx) <-
+    judgmentContext
+  let tx = stAnnTx ^. txStAnnTxG
+      legacyMode = stAnnTx ^. plutusLegacyModeStAnnTxG
       withdrawals = tx ^. bodyTxL . withdrawalsTxBodyL
       accounts = certState ^. certDStateL . accountsL
+      certsEnv = Conway.CertsEnv tx pp curEpoch committee committeeProposals
 
   network <- liftSTS $ asks networkId
 
+  checkAccountBalanceIntervals
+    network
+    accounts
+    (tx ^. bodyTxL . accountBalanceIntervalsTxBodyL)
+    AccountBalanceIntervalsWrongNetwork
+    AccountBalanceIntervalsUnknownAccounts
+    AccountBalanceIntervalsMismatch
+  checkAccountBalanceIntervals
+    network
+    originalAccounts
+    (tx ^. bodyTxL . startingAccountBalanceIntervalsTxBodyL)
+    StartingAccountBalanceIntervalsWrongNetwork
+    StartingAccountBalanceIntervalsUnknownAccounts
+    StartingAccountBalanceIntervalsMismatch
+
   runTest $ Shelley.validateWrongNetworkWithdrawal network (tx ^. bodyTxL)
   runTest $ validateWrongNetworkInDirectDeposit network (tx ^. bodyTxL)
-
   validateWithdrawals legacyMode network withdrawals accounts
 
   let certStateBeforeCerts =
@@ -227,7 +300,8 @@ dijkstraEntitiesTransition = do
           & Conway.updateVotingDRepExpiries tx curEpoch (pp ^. ppDRepActivityL)
           & certDStateL . accountsL %~ applyWithdrawals withdrawals
   certStateAfterCerts <-
-    trans @(EraRule "CERTS" era) $ TRC (certsEnv, certStateBeforeCerts, certificates)
+    trans @(EraRule "CERTS" era) $
+      TRC (certsEnv, certStateBeforeCerts, StrictSeq.fromStrict $ tx ^. bodyTxL . certsTxBodyL)
 
   let directDeposits = tx ^. bodyTxL . directDepositsTxBodyL
       accountsAfterCerts = certStateAfterCerts ^. certDStateL . accountsL
@@ -327,3 +401,57 @@ instance
   where
   wrapFailed = CertsFailure
   wrapEvent = CertsEvent
+
+accountBalanceIntervalContains :: Coin -> AccountBalanceInterval era -> Bool
+accountBalanceIntervalContains bal = \case
+  AccountBalanceLowerBound (Inclusive lo) -> lo <= bal
+  AccountBalanceUpperBound (Exclusive hi) -> bal < hi
+  AccountBalanceBothBounds (Inclusive lo) (Exclusive hi) -> lo <= bal && bal < hi
+  AccountBalanceExact n -> bal == n
+
+accountBalanceIntervalsMismatches ::
+  EraAccounts era =>
+  AccountBalanceIntervals era ->
+  Accounts era ->
+  Map.Map AccountAddress (AccountBalanceInterval era)
+accountBalanceIntervalsMismatches (AccountBalanceIntervals intervals) accounts =
+  Map.filterWithKey
+    (\acct interval -> not (accountBalanceIntervalContains (balanceOf acct) interval))
+    intervals
+  where
+    balanceOf acct =
+      maybe
+        mempty
+        (fromCompact . (^. balanceAccountStateL))
+        (lookupAccountState (acct ^. accountAddressCredentialL) accounts)
+
+accountBalanceIntervalsMissingAccounts ::
+  EraAccounts era =>
+  AccountBalanceIntervals era ->
+  Accounts era ->
+  Map.Map AccountAddress (AccountBalanceInterval era)
+accountBalanceIntervalsMissingAccounts (AccountBalanceIntervals intervals) accounts =
+  Map.filterWithKey
+    (\acct _ -> not (isAccountRegistered (acct ^. accountAddressCredentialL) accounts))
+    intervals
+
+accountBalanceIntervalsWrongNetwork ::
+  AccountBalanceIntervals era ->
+  Network ->
+  Set AccountAddress
+accountBalanceIntervalsWrongNetwork (AccountBalanceIntervals intervals) network =
+  Map.keysSet $ Map.filterWithKey (\a _ -> aaNetworkId a /= network) intervals
+
+checkAccountBalanceIntervals ::
+  EraAccounts era =>
+  Network ->
+  Accounts era ->
+  AccountBalanceIntervals era ->
+  (Network -> NonEmptySet AccountAddress -> PredicateFailure sts) ->
+  (NonEmptyMap AccountAddress (AccountBalanceInterval era) -> PredicateFailure sts) -> -- unknownaccounts
+  (NonEmptyMap AccountAddress (AccountBalanceInterval era) -> PredicateFailure sts) -> -- mismatch
+  Rule sts ctx ()
+checkAccountBalanceIntervals network accounts intervals mkNetwork mkUnknownAccounts mkMismatch = do
+  failOnNonEmptySet (accountBalanceIntervalsWrongNetwork intervals network) $ mkNetwork network
+  failOnNonEmptyMap (accountBalanceIntervalsMissingAccounts intervals accounts) mkUnknownAccounts
+  failOnNonEmptyMap (accountBalanceIntervalsMismatches intervals accounts) mkMismatch